refactor(store): flatten codec/ into codec.rs
Collapse codec/{owning, borrowing}.rs and the codec/serde/ subdirectory
into a single codec.rs at the crate root. Codec and BorrowingCodec
traits live at the top; the feature-gated serde adapter (SerdeFormat,
SerdeCodec) sits in an inline `pub mod serde` block, with Json and
JsonCodec in a nested `pub mod json` inside that.

Public paths unchanged:
- nexus_store::codec::{Codec, BorrowingCodec}
- nexus_store::codec::serde::{SerdeCodec, SerdeFormat}
- nexus_store::codec::serde::json::{Json, JsonCodec}
