--- crates/nexus-store/src/codec/mod.rs
@@ -1,9 +1,0 @@
-mod borrowing;
-mod owning;
-#[cfg(feature = "serde")]
-pub mod serde;
-
-#[cfg(feature = "serde")]
-pub use self::serde::{SerdeCodec, SerdeFormat};
-pub use borrowing::BorrowingCodec;
-pub use owning::Codec;
